Don't call `webView.reload()` if the webView has no current URL
The functionality added in https://github.com/hotwired/turbo-android/pull/138 (as well as my workaround in https://github.com/hotwired/turbo-android/issues/137) had a bug: if the webView has no URL loaded, calling `refresh` will result in a spinner appearing forever, and no requests being made to the server.

To replicate this:

- Add a call to `delegate.refresh()` in `onResume`
- Force close the app
- Open the app

Our real life example was that we had a call to `refresh()` in `onResume` that ran if the app had not been open for a while, to ensure fresh content. (Is there a better approach to this? How does Basecamp/Hey ensure fresh content on app load - either because the app was in memory in the background for a while or because Turbo/the Webview has cached the last screen?)

The fix is to not call `webView.reload()` if `webView.url` is null. Instead just load the URL as normal.