Call super in `shouldNavigateTo` override
Since https://github.com/hotwired/turbo-android/pull/163, the base implementation of `shouldNavigateTo` is useful because it checks for invalid URLs. So if you override this method we should still recommend you call `super` to get the base implementation.

While here I turned the `when` into an `if` per [this](https://kotlinlang.org/docs/coding-conventions.html#if-versus-when), but I'm certainly not a Kotlin expert and happy to change it back if you prefer.